test: Get test working
