+++ config/storybook/start/addons.js
@@ -1,0 +1,1 @@
+import '@storybook/addon-knobs/register';
