+++ config/storybook/start/webpack.config.js
@@ -1,0 +1,3 @@
+const storybookWebpackConfig = require('../storybookWebpackConfig');
+
+module.exports = storybookWebpackConfig();
