+++ test/test-cases/configure-test/sku.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   target: 'foo/bar',
+  storybookTarget: 'storybook/foobar',
   dangerouslySetESLintConfig: config => {
     config.rules = config.rules || {};
     config.rules['no-console'] = 0;
