+++ test/test-cases/custom-src-paths/.gitignore
@@ -2,6 +2,7 @@
 .eslintrc
 .prettierrc
 coverage/
+dist-storybook/
 dist/
 report/
 # end managed by sku
