+++ test/test-cases/public-path/.eslintignore
@@ -1,5 +1,6 @@
 # managed by sku
 coverage/
+dist-storybook/
 dist/static/
 report/
 # end managed by sku
