+++ test/test-cases/public-path/.gitignore
@@ -2,6 +2,7 @@
 .eslintrc
 .prettierrc
 coverage/
+dist-storybook/
 dist/static/
 report/
 # end managed by sku
