+++ test/test-cases/typescript-css-modules/app/.prettierignore
@@ -1,5 +1,6 @@
 # managed by sku
 coverage/
+dist-storybook/
 dist/
 report/
 # end managed by sku
